Organize runtime value function families
- Separate value-return runners under their runtime return-family owners.
- Co-locate Int and Bool graph-host protocols with their exact tests.
