+++ src/runtime/function/value.rs
@@ -1,579 +1,21 @@
-use super::{evaluate, run_tail};
-use crate::plan::execution::HostedExecution;
-use crate::plan::execution::function::{
-    BitArrayFunctionId, BoolFunctionBody, BoolFunctionId, CustomFunctionId, ExecutableFunction,
-    FloatFunctionId, IntFunctionBody, IntFunctionId, NeverFunctionId, NilFunctionId,
-    StringFunctionId, TupleFunctionId, UtfCodepointFunctionId, ValueFunctionEntry,
-};
-use crate::plan::execution::graph::ParamLocal;
-use crate::plan::execution::host::{HostBoolFunctionId, HostIntFunctionId};
-use crate::plan::execution::runtime::RuntimeExecutionPlan;
-use crate::runtime::ExecutableRuntimePlan;
-use crate::runtime::error::ExecutionResult;
-use crate::runtime::evaluated::{
-    EvaluatedBitArray, EvaluatedCustomValue, EvaluatedNeverFunction, EvaluatedValue,
-};
-use crate::runtime::graph::RetainedValues;
-use crate::runtime::state::RuntimeState;
-use ecow::EcoString;
-use num_bigint::BigInt;
-use std::convert::Infallible;
+mod bit_array;
+mod bool;
+mod custom;
+mod float;
+mod int;
+mod never;
+mod nil;
+mod string;
+mod tuple;
+mod utf_codepoint;
 
-pub(in crate::runtime) trait RuntimeIntFunction<Plan: RuntimeExecutionPlan> {
-    fn evaluate<EvaluateGraph>(
-        &self,
-        plan: &Plan,
-        state: &mut RuntimeState,
-        inputs: RetainedValues,
-        evaluate_graph: EvaluateGraph,
-    ) -> ExecutionResult<super::EvaluatedFunctionExit<BigInt, IntFunctionId>>
-    where
-        EvaluateGraph: FnOnce(
-            &Plan,
-            &mut RuntimeState,
-            &ExecutableFunction<IntFunctionBody>,
-            RetainedValues,
-        ) -> ExecutionResult<
-            super::EvaluatedFunctionExit<BigInt, IntFunctionId>,
-        >;
-
-    fn parameter_locals(&self, plan: &Plan) -> Vec<ParamLocal>;
-}
-
-impl<Plan: RuntimeExecutionPlan> RuntimeIntFunction<Plan> for ExecutableFunction<IntFunctionBody> {
-    fn evaluate<EvaluateGraph>(
-        &self,
-        plan: &Plan,
-        state: &mut RuntimeState,
-        inputs: RetainedValues,
-        evaluate_graph: EvaluateGraph,
-    ) -> ExecutionResult<super::EvaluatedFunctionExit<BigInt, IntFunctionId>>
-    where
-        EvaluateGraph: FnOnce(
-            &Plan,
-            &mut RuntimeState,
-            &ExecutableFunction<IntFunctionBody>,
-            RetainedValues,
-        ) -> ExecutionResult<
-            super::EvaluatedFunctionExit<BigInt, IntFunctionId>,
-        >,
-    {
-        evaluate_graph(plan, state, self, inputs)
-    }
-
-    fn parameter_locals(&self, _plan: &Plan) -> Vec<ParamLocal> {
-        self.entry()
-            .params(self.body())
-            .iter()
-            .map(|slot| slot.local().clone())
-            .collect()
-    }
-}
-
-impl RuntimeIntFunction<HostedExecution>
-    for ValueFunctionEntry<IntFunctionBody, HostIntFunctionId>
-{
-    fn evaluate<EvaluateGraph>(
-        &self,
-        plan: &HostedExecution,
-        state: &mut RuntimeState,
-        inputs: RetainedValues,
-        evaluate_graph: EvaluateGraph,
-    ) -> ExecutionResult<super::EvaluatedFunctionExit<BigInt, IntFunctionId>>
-    where
-        EvaluateGraph: FnOnce(
-            &HostedExecution,
-            &mut RuntimeState,
-            &ExecutableFunction<IntFunctionBody>,
-            RetainedValues,
-        ) -> ExecutionResult<
-            super::EvaluatedFunctionExit<BigInt, IntFunctionId>,
-        >,
-    {
-        match self {
-            ValueFunctionEntry::Graph(function) => evaluate_graph(plan, state, function, inputs),
-            ValueFunctionEntry::Host(target) => Ok(super::EvaluatedFunctionExit::Return(
-                plan.host_int_function(*target).call(&inputs),
-            )),
-        }
-    }
-
-    fn parameter_locals(&self, plan: &HostedExecution) -> Vec<ParamLocal> {
-        match self {
-            ValueFunctionEntry::Graph(function) => function.parameter_locals(plan),
-            ValueFunctionEntry::Host(target) => {
-                plan.host_int_function(*target).parameters().to_vec()
-            }
-        }
-    }
-}
-
-pub(in crate::runtime) trait RuntimeBoolFunction<Plan: RuntimeExecutionPlan> {
-    fn evaluate<EvaluateGraph>(
-        &self,
-        plan: &Plan,
-        state: &mut RuntimeState,
-        inputs: RetainedValues,
-        evaluate_graph: EvaluateGraph,
-    ) -> ExecutionResult<super::EvaluatedFunctionExit<bool, BoolFunctionId>>
-    where
-        EvaluateGraph: FnOnce(
-            &Plan,
-            &mut RuntimeState,
-            &ExecutableFunction<BoolFunctionBody>,
-            RetainedValues,
-        ) -> ExecutionResult<
-            super::EvaluatedFunctionExit<bool, BoolFunctionId>,
-        >;
-
-    fn parameter_locals(&self, plan: &Plan) -> Vec<ParamLocal>;
-}
-
-impl<Plan: RuntimeExecutionPlan> RuntimeBoolFunction<Plan>
-    for ExecutableFunction<BoolFunctionBody>
-{
-    fn evaluate<EvaluateGraph>(
-        &self,
-        plan: &Plan,
-        state: &mut RuntimeState,
-        inputs: RetainedValues,
-        evaluate_graph: EvaluateGraph,
-    ) -> ExecutionResult<super::EvaluatedFunctionExit<bool, BoolFunctionId>>
-    where
-        EvaluateGraph: FnOnce(
-            &Plan,
-            &mut RuntimeState,
-            &ExecutableFunction<BoolFunctionBody>,
-            RetainedValues,
-        ) -> ExecutionResult<
-            super::EvaluatedFunctionExit<bool, BoolFunctionId>,
-        >,
-    {
-        evaluate_graph(plan, state, self, inputs)
-    }
-
-    fn parameter_locals(&self, _plan: &Plan) -> Vec<ParamLocal> {
-        self.entry()
-            .params(self.body())
-            .iter()
-            .map(|slot| slot.local().clone())
-            .collect()
-    }
-}
-
-impl RuntimeBoolFunction<HostedExecution>
-    for ValueFunctionEntry<BoolFunctionBody, HostBoolFunctionId>
-{
-    fn evaluate<EvaluateGraph>(
-        &self,
-        plan: &HostedExecution,
-        state: &mut RuntimeState,
-        inputs: RetainedValues,
-        evaluate_graph: EvaluateGraph,
-    ) -> ExecutionResult<super::EvaluatedFunctionExit<bool, BoolFunctionId>>
-    where
-        EvaluateGraph: FnOnce(
-            &HostedExecution,
-            &mut RuntimeState,
-            &ExecutableFunction<BoolFunctionBody>,
-            RetainedValues,
-        ) -> ExecutionResult<
-            super::EvaluatedFunctionExit<bool, BoolFunctionId>,
-        >,
-    {
-        match self {
-            ValueFunctionEntry::Graph(function) => evaluate_graph(plan, state, function, inputs),
-            ValueFunctionEntry::Host(target) => Ok(super::EvaluatedFunctionExit::Return(
-                plan.host_bool_function(*target).call(&inputs),
-            )),
-        }
-    }
-
-    fn parameter_locals(&self, plan: &HostedExecution) -> Vec<ParamLocal> {
-        match self {
-            ValueFunctionEntry::Graph(function) => function.parameter_locals(plan),
-            ValueFunctionEntry::Host(target) => {
-                plan.host_bool_function(*target).parameters().to_vec()
-            }
-        }
-    }
-}
-
-pub(in crate::runtime) fn run_never(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    function: NeverFunctionId,
-    inputs: RetainedValues,
-) -> ExecutionResult<Infallible> {
-    run_tail(
-        plan,
-        state,
-        function,
-        inputs,
-        |plan, state, function, inputs| {
-            evaluate(plan, state, plan.never_function(*function).body(), inputs)
-        },
-        |_, _, target| target,
-    )
-}
-
-pub(in crate::runtime) fn run_int(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    mut function: IntFunctionId,
-    mut inputs: RetainedValues,
-) -> ExecutionResult<BigInt> {
-    loop {
-        match plan.int_function(function).evaluate(
-            plan,
-            state,
-            inputs,
-            |plan, state, function, inputs| evaluate(plan, state, function.body(), inputs),
-        )? {
-            super::EvaluatedFunctionExit::Return(value) => return Ok(value),
-            super::EvaluatedFunctionExit::TailCall {
-                function: target,
-                args,
-            } => {
-                function = target;
-                inputs = args;
-            }
-        }
-    }
-}
-
-pub(in crate::runtime) fn run_float(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    function: FloatFunctionId,
-    inputs: RetainedValues,
-) -> ExecutionResult<f64> {
-    run_tail(
-        plan,
-        state,
-        function,
-        inputs,
-        |plan, state, function, inputs| {
-            evaluate(plan, state, plan.float_function(*function).body(), inputs)
-        },
-        |_, _, target| target,
-    )
-}
-
-pub(in crate::runtime) fn run_string(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    function: StringFunctionId,
-    inputs: RetainedValues,
-) -> ExecutionResult<EcoString> {
-    run_tail(
-        plan,
-        state,
-        function,
-        inputs,
-        |plan, state, function, inputs| {
-            evaluate(plan, state, plan.string_function(*function).body(), inputs)
-        },
-        |_, _, target| target,
-    )
-}
-
-pub(in crate::runtime) fn run_bit_array(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    function: BitArrayFunctionId,
-    inputs: RetainedValues,
-) -> ExecutionResult<EvaluatedBitArray> {
-    run_tail(
-        plan,
-        state,
-        function,
-        inputs,
-        |plan, state, function, inputs| {
-            evaluate(
-                plan,
-                state,
-                plan.bit_array_function(*function).body(),
-                inputs,
-            )
-        },
-        |_, _, target| target,
-    )
-}
-
-pub(in crate::runtime) fn run_utf_codepoint(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    function: UtfCodepointFunctionId,
-    inputs: RetainedValues,
-) -> ExecutionResult<char> {
-    run_tail(
-        plan,
-        state,
-        function,
-        inputs,
-        |plan, state, function, inputs| {
-            evaluate(
-                plan,
-                state,
-                plan.utf_codepoint_function(*function).body(),
-                inputs,
-            )
-        },
-        |_, _, target| target,
-    )
-}
-
-pub(in crate::runtime) fn run_custom(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    function: CustomFunctionId,
-    inputs: RetainedValues,
-) -> ExecutionResult<EvaluatedCustomValue> {
-    run_tail(
-        plan,
-        state,
-        function,
-        inputs,
-        |plan, state, function, inputs| {
-            evaluate(
-                plan,
-                state,
-                plan.custom_function(*function).body().function_body(),
-                inputs,
-            )
-        },
-        |plan, function, target| plan.custom_function(*function).body().function_id(target),
-    )
-}
-
-pub(in crate::runtime) fn run_bool(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    mut function: BoolFunctionId,
-    mut inputs: RetainedValues,
-) -> ExecutionResult<bool> {
-    loop {
-        match plan.bool_function(function).evaluate(
-            plan,
-            state,
-            inputs,
-            |plan, state, function, inputs| evaluate(plan, state, function.body(), inputs),
-        )? {
-            super::EvaluatedFunctionExit::Return(value) => return Ok(value),
-            super::EvaluatedFunctionExit::TailCall {
-                function: target,
-                args,
-            } => {
-                function = target;
-                inputs = args;
-            }
-        }
-    }
-}
-
-pub(in crate::runtime) fn run_nil(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    function: NilFunctionId,
-    inputs: RetainedValues,
-) -> ExecutionResult<()> {
-    run_tail(
-        plan,
-        state,
-        function,
-        inputs,
-        |plan, state, function, inputs| {
-            evaluate(plan, state, plan.nil_function(*function).body(), inputs)
-        },
-        |_, _, target| target,
-    )
-}
-
-pub(in crate::runtime) fn run_tuple(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    function: TupleFunctionId,
-    inputs: RetainedValues,
-) -> ExecutionResult<Vec<EvaluatedValue>> {
-    run_tail(
-        plan,
-        state,
-        function,
-        inputs,
-        |plan, state, function, inputs| {
-            evaluate(plan, state, plan.tuple_function(*function).body(), inputs)
-        },
-        |_, _, target| target,
-    )
-}
-
-pub(in crate::runtime) fn run_never_value(
-    plan: &impl ExecutableRuntimePlan,
-    state: &mut RuntimeState,
-    function: EvaluatedNeverFunction,
-    mut inputs: RetainedValues,
-) -> ExecutionResult<Infallible> {
-    inputs.append_captures(function.captures());
-    run_never(plan, state, function.runtime_id(), inputs)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{RuntimeBoolFunction as _, RuntimeIntFunction as _};
-    use crate::plan::execution::function::{BoolFunctionId, IntFunctionId};
-    use crate::plan::execution::graph::{BoolLocalId, IntLocalId, ParamLocal};
-    use crate::plan::execution::runtime::RuntimeExecutionPlan;
-    use crate::{
-        HostModule, HostModules, HostedExecution, ModuleSource, PackageSource, Value,
-        compile_typed_host_program, compile_typed_module, plan_host_program, plan_module, run_main,
-    };
-    use num_bigint::BigInt;
-
-    #[test]
-    fn plain_int_function_protocol_executes_graph_entries() {
-        let source = r#"
-fn increment(value: Int) {
-  value + 1
-}
-
-pub fn main() {
-  increment(41)
-}
-"#;
-        let typed =
-            compile_typed_module("main", "main.gleam", source).expect("source should compile");
-        let plan = plan_module(typed).expect("source should plan");
-        let execution = crate::ExecutionPlan::from_module_plan(plan);
-        let function = RuntimeExecutionPlan::int_function(&execution, IntFunctionId(1));
-
-        assert_eq!(
-            function.parameter_locals(&execution),
-            [ParamLocal::Int(IntLocalId(0))],
-        );
-        assert_eq!(
-            run_main(&execution, &mut Vec::new()),
-            Ok(Value::Int(42.into()))
-        );
-    }
-
-    #[test]
-    fn plain_bool_function_protocol_executes_graph_entries() {
-        let source = r#"
-fn identity(value: Bool) {
-  value
-}
-
-pub fn main() {
-  identity(True)
-}
-"#;
-        let typed =
-            compile_typed_module("main", "main.gleam", source).expect("source should compile");
-        let plan = plan_module(typed).expect("source should plan");
-        let execution = crate::ExecutionPlan::from_module_plan(plan);
-        let function = RuntimeExecutionPlan::bool_function(&execution, BoolFunctionId(1));
-
-        assert_eq!(
-            function.parameter_locals(&execution),
-            [ParamLocal::Bool(BoolLocalId(0))],
-        );
-        assert_eq!(run_main(&execution, &mut Vec::new()), Ok(Value::Bool(true)),);
-    }
-
-    #[test]
-    fn hosted_int_function_protocol_executes_graph_and_host_entries() {
-        let math = HostModule::new("host_support", "host/math")
-            .expect("host module should be valid")
-            .with_function("add", <BigInt as std::ops::Add>::add)
-            .expect("host function should be valid");
-        let hosts = HostModules::new([math]).expect("host modules should be unique");
-        let source = r#"
-import host/math
-
-fn increment(value: Int) {
-  value + 1
-}
-
-pub fn main() {
-  increment(math.add(20, 21))
-}
-"#;
-        let typed = compile_typed_host_program(
-            "application",
-            "main",
-            [PackageSource::new(
-                "application",
-                ["host_support"],
-                [ModuleSource::new("main", "main.gleam", source)],
-            )],
-            hosts,
-        )
-        .expect("host source should compile");
-        let plan = plan_host_program(typed).expect("host source should plan");
-        let execution = HostedExecution::from_module_plan(plan);
-        let host = RuntimeExecutionPlan::int_function(&execution, IntFunctionId(1));
-        let graph = RuntimeExecutionPlan::int_function(&execution, IntFunctionId(2));
-
-        assert_eq!(
-            graph.parameter_locals(&execution),
-            [ParamLocal::Int(IntLocalId(0))],
-        );
-        assert_eq!(
-            host.parameter_locals(&execution),
-            [
-                ParamLocal::Int(IntLocalId(0)),
-                ParamLocal::Int(IntLocalId(1)),
-            ],
-        );
-        assert_eq!(
-            execution.run_main(&mut Vec::new()),
-            Ok(Value::Int(42.into())),
-        );
-    }
-
-    #[test]
-    fn hosted_bool_function_protocol_executes_graph_and_host_entries() {
-        let predicates = HostModule::new("host_support", "host/predicates")
-            .expect("host module should be valid")
-            .with_function("is_positive", |value: BigInt| value > BigInt::from(0))
-            .expect("host function should be valid");
-        let hosts = HostModules::new([predicates]).expect("host modules should be unique");
-        let source = r#"
-import host/predicates
-
-fn identity(value: Bool) {
-  value
-}
-
-pub fn main() {
-  identity(predicates.is_positive(1))
-}
-"#;
-        let typed = compile_typed_host_program(
-            "application",
-            "main",
-            [PackageSource::new(
-                "application",
-                ["host_support"],
-                [ModuleSource::new("main", "main.gleam", source)],
-            )],
-            hosts,
-        )
-        .expect("host source should compile");
-        let plan = plan_host_program(typed).expect("host source should plan");
-        let execution = HostedExecution::from_module_plan(plan);
-        let host = RuntimeExecutionPlan::bool_function(&execution, BoolFunctionId(1));
-        let graph = RuntimeExecutionPlan::bool_function(&execution, BoolFunctionId(2));
-
-        assert_eq!(
-            graph.parameter_locals(&execution),
-            [ParamLocal::Bool(BoolLocalId(0))],
-        );
-        assert_eq!(
-            host.parameter_locals(&execution),
-            [ParamLocal::Int(IntLocalId(0))],
-        );
-        assert_eq!(execution.run_main(&mut Vec::new()), Ok(Value::Bool(true)),);
-    }
-}
+pub(in crate::runtime) use bit_array::run_bit_array;
+pub(in crate::runtime) use bool::{RuntimeBoolFunction, run_bool};
+pub(in crate::runtime) use custom::run_custom;
+pub(in crate::runtime) use float::run_float;
+pub(in crate::runtime) use int::{RuntimeIntFunction, run_int};
+pub(in crate::runtime) use never::{run_never, run_never_value};
+pub(in crate::runtime) use nil::run_nil;
+pub(in crate::runtime) use string::run_string;
+pub(in crate::runtime) use tuple::run_tuple;
+pub(in crate::runtime) use utf_codepoint::run_utf_codepoint;
